+++ changelog.md
@@ -4,6 +4,42 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# 3.0.0
+### Added
+- `fromArrayBuffer` now supports trimming empty leading layers.
+  - ONBS automatically generates extra layers past the last populated layer.
+- Created basic testing via `jest`.
+- Added ESLint and plugins to `devDependencies`.
+- Added `test` script which lints and runs tests.
+- The `options` argument in `fromArrayBuffer`.
+- The `FromArrayBufferOptions` interface.
+
+### Changed
+- The `build` now runs the `test` script.
+- All interfaces, classes, and objects no longer have a default export.
+  - This should not affect `@encode42/nbs.js` imports.
+- All interfaces have been moved to their respective class.
+  - For example, `InstrumentMeta.ts` -> `Instrument.ts`.
+
+### Removed
+- The `nbs/interfaces` directory.
+- The `SongMeta.ts` interface file.
+  - Moved to `Song.ts`.
+- The `SongStats.ts` interface file.
+  - Moved to `Song.ts`.
+- The `SongAutosaveOptions.ts` interface file.
+  - Moved to `Song.ts`.
+- The `SongLoopOptions.ts` interface file.
+  - Moved to `Song.ts`.
+- The `LayerMeta.ts` interface file.
+  - Moved to `Layer.ts`.
+- The `NoteOptions.ts` interface file.
+  - Moved to `Note.ts`.
+- The `InstrumentMeta.ts` interface file.
+  - Moved to `Instrument.ts`.
+- The `InstrumentOptions.ts` interface file.
+  - Moved to `InstrumentOptions.ts`.
+
 # 2.1.2
 ### Changed
 - Updated all `devDependencies`.